requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
aiofiles
databases
aiosqlite
python-multipart